Conversational_Chatbot/requirements.txt
langchain>=1.0,<2
langchain-core>=0.3,<2
langchain-google-genai>=2.0
python-dotenv>=1.0
streamlit>=1.28
